+++ src/main/java/jp/thotta/ifinance/collector/news/CompanyNewsCollector3668.java
@@ -12,6 +12,7 @@ import org.jsoup.select.Elements;
 import org.jsoup.nodes.Element;
 
 import jp.thotta.ifinance.collector.CompanyNewsCollector;
+import jp.thotta.ifinance.collector.BaseCompanyNewsCollector;
 import jp.thotta.ifinance.model.CompanyNews;
 import jp.thotta.ifinance.common.MyDate;
 import jp.thotta.ifinance.common.Scraper;
@@ -24,7 +25,7 @@ import jp.thotta.ifinance.common.ParseNewsPageException;
  * @author toru1055
  */
 public class CompanyNewsCollector3668
-  extends AbstractCompanyNewsCollector
+  extends BaseCompanyNewsCollector
   implements CompanyNewsCollector {
   private static final int stockId = 3668;
   private static final String PR_URL = "http://colopl.co.jp/news/";
